% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alpha.R
\name{add_alpha}
\alias{add_alpha}
\title{Add transparency to supplied vector of colours}
\usage{
add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{vector of colours}

\item{alpha}{a value in (0, 1]. Default is 1 which does not change the transparency. 0 would be an invisible colour.}
}
\description{
Add transparency to supplied vector of colours
}
\details{
The returned colours are hex colour values even if colours were
entered as names. e.g., \code{add_alpha("black", 0.5)} returns \code{"#00000080"}
}
\examples{
add_alpha(c("black", "red", "yellow", "blue"), 0.5)


}
\author{
Ty Stanford <tystan@gmail.com>
}
