% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_rowwise.R
\name{closure_rowwise}
\alias{closure_rowwise}
\title{Apply closure operator over rows}
\usage{
closure_rowwise(y, clo_val = 1)
}
\arguments{
\item{y}{\code{data.frame} or \code{matrix} with of \code{n} observations (rows) and \code{D} compositional components (columns)}

\item{clo_val}{a positive value for each row to be closed to (i.e., each rows sums to this)}
}
\description{
Apply closure operator over rows
}
\examples{
(c_dat1 <- matrix(c(rep(0.25, 4), 1:4 / 10, 1:4), byrow = TRUE, ncol = 4))
closure_rowwise(c_dat1)

}
\author{
Ty Stanford <tystan@gmail.com>
}
