% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_inverse_sqrt.R
\name{covar_inverse_sqrt}
\alias{covar_inverse_sqrt}
\title{Calculate the inverse square-root of the variance-covariance matrix}
\usage{
covar_inverse_sqrt(covar_mat, verbose = FALSE)
}
\arguments{
\item{covar_mat}{a positive-definite, square (n by n) matrix}

\item{verbose}{provide working (FALSE, default)}
}
\value{
Find the matrix U such that

\code{inverse(covar_mat)} \eqn{= U U}{= U U}
}
\description{
Calculate the inverse square-root of the variance-covariance matrix
}
\note{
Uses \code{expm::sqrtm()} so should be optimised for speed over 
eigen decompositions manually, i.e., the implementation in 
\code{covar_to_power(., pow = -1/2)}
}
\examples{
A <- diag(2)
covar_inverse_sqrt(A)
covar_inverse_sqrt(A, verbose = TRUE)
covar_to_power(A, pow = -1/2, verbose = TRUE)
all(abs(covar_inverse_sqrt(A) \%*\% covar_inverse_sqrt(A) - solve(A)) < 1e-12)
all(abs(covar_to_power(A, pow = -1/2) \%*\% covar_to_power(A, pow = -1/2) - solve(A)) < 1e-12)

B <- diag(c(1, 4))
covar_inverse_sqrt(B)
covar_inverse_sqrt(B, verbose = TRUE)
covar_to_power(B, pow = -1/2, verbose = TRUE)
all(abs(covar_inverse_sqrt(B) \%*\% covar_inverse_sqrt(B) - solve(B)) < 1e-12)
all(abs(covar_to_power(B, pow = -1/2) \%*\% covar_to_power(B, pow = -1/2) - solve(B)) < 1e-12)

C <- matrix(c(4, -2, -2, 4), nrow = 2)
covar_inverse_sqrt(C)
covar_inverse_sqrt(C, verbose = TRUE)
covar_to_power(C, pow = -1/2, verbose = TRUE)
all(abs(covar_inverse_sqrt(C) \%*\% covar_inverse_sqrt(C) - solve(C)) < 1e-12)
all(abs(covar_to_power(C, pow = -1/2) \%*\% covar_to_power(C, pow = -1/2) - solve(C)) < 1e-12)

}
