% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vt.R
\name{create_vt}
\alias{create_vt}
\title{Create the transpose of a V matrix (i.e., V^T)}
\usage{
create_vt(n_comp, normalise = TRUE)
}
\description{
Create the transpose of a V matrix (i.e., V^T)
}
\details{
Note that:

\code{ilr(x) = ln(x) \%*\% t(V)}

\code{-> x = clo(exp(ilr(x) \%*\% V))}

as

\code{t(V) \%*\% V =} \eqn{I_{D-1}}, and

\code{V \%*\% t(V) =} \eqn{I_{D} - \frac{1}{D} \times 1_{DxD}}
}
\examples{
create_vt(4)
(VT <- create_vt(3))
VT \%*\% t(VT) 
t(VT) \%*\% VT  
}
