% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean_colwise.R
\name{geo_mean_colwise}
\alias{geo_mean_colwise}
\title{Calculate the geomtric mean down columns}
\usage{
geo_mean_colwise(x)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix} with of \code{n} observations (rows) and \code{D} compositional components (columns)}
}
\description{
Calculate the geomtric mean down columns in mtrix/data.frame
}
\examples{
(c_dat1 <- matrix(c(rep(0.25, 4), 1:4 / 10, 1:4), byrow = TRUE, ncol = 4))
geo_mean_colwise(c_dat1)

}
\author{
Ty Stanford <tystan@gmail.com>
}
