% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viri_scale.R
\name{get_viri_scale}
\alias{get_viri_scale}
\title{Create a colour scale/palette based on the viridis-D scale}
\usage{
get_viri_scale(n = 101, alpha = 1)
}
\arguments{
\item{n}{number of colours in the palette (default \code{101})}

\item{alpha}{transparency of the returned colours (default \code{1})}
}
\description{
Create a colour scale/palette based on the viridis-D scale
}
\details{
Returns a vector of length \code{n} with hex colour values that can be used for plotting.
}
\examples{
get_viri_scale(4)
get_viri_scale(4, 0.5) # see-through

}
\author{
Ty Stanford <tystan@gmail.com>
}
