% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_convex.R
\name{is_convex_over_simplex}
\alias{is_convex_over_simplex}
\title{Determine if a suppled function is convex}
\usage{
is_convex_over_simplex(func, x_dim, n_rep = 10000, seed_ = NULL)
}
\arguments{
\item{func}{a function with a single input that is a vector that is of dimension \code{x_dim} (see below)}

\item{x_dim}{the dimension of the input required}

\item{n_rep}{(default = 10000) the number of randomly generated \code{(x,y,a,b)}-tuples generated to
assess \code{f(ax + by) <= a f(x) + b f(y)}}

\item{seed_}{(default: NULL) a seed for the random number generation. If NULL, no seed set and is non-reproducible}
}
\description{
Function to evaluate whether \code{f(ax + by) <= a f(x) + b f(y)} is true for all
 \code{x,y} in a function's domain and
}
\details{
# Function to evaluate whether \code{f(ax + by) <= a f(x) + b f(y)} is true for all \code{x,y} in a function's domain and
  \code{a + b = 1} with \code{a,b >= 0}. i.e., determine if a suppled \code{f()} is convex.
The returned value is an approximation in that not all x,y,a,b are enumerated. Confidence in a "true"
  result increases as the number of repititions increases.

Either \code{TRUE} or \code{FALSE} is returned. Additionally details of is_convex's findings is printed to console.

It is assumed that the input of the function is a \code{x_dim}-dimensional composition
(that is sums to 1 and elements are non-negative)
Only a returned value of \code{FALSE} can be consided a difinitive answer to whether convexity exists as a counter example
  of \code{f(ax + by) <= a f(x) + b f(y)} has been found (analogous to proof by contradiction)
If \code{TRUE} is returned it just means no counter examples have been found
}
\examples{
example_fun_1 <- function(x) {
 -sum(log(x))
}
# 2-D check, takes ~10 sec (to reduce time use n_rep = 1e4 or similar)
is_convex_over_simplex(example_fun_1, 2, n_rep = 1e4, seed_ = 1234)
# 4-D check
is_convex_over_simplex(example_fun_1, 4, n_rep = 1e3, seed_ = 1234)
}
\author{
Ty Stanford <tystan@gmail.com>
}
