% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cts_to_scale.R
\name{map_cts_to_scale}
\alias{map_cts_to_scale}
\title{Map numeric values to corresponding viridis-D scale palette}
\usage{
map_cts_to_scale(x, alpha = 0.9)
}
\arguments{
\item{x}{a vector of numeric values}

\item{alpha}{transparency of the returned colours (default \code{0.9})}
}
\description{
Map numeric values to corresponding viridis-D scale palette
}
\details{
Returns a vector of equal size of \code{x} with viridis-D scale hex colour 
values (with the min and max values of \code{x} defining the range of the palette).
}
\examples{
set.seed(1234)
(x <- rnorm(10))
x_scaled <- scale01(x)
x_cols <- map_cts_to_scale(x, alpha = 0.8)
plot(x, x_scaled, col = x_cols, pch = 16, cex = 3)
}
\author{
Ty Stanford <tystan@gmail.com>
}
