% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_simplex_grid.R
\name{mk_simplex_grid}
\alias{mk_simplex_grid}
\title{Create a grid on a m-dim simplex}
\usage{
mk_simplex_grid(dim, step_size = 0.1)
}
\arguments{
\item{dim}{the dimension of the simplex. This will be the number of columns in the returned data}

\item{step_size}{(default 0.1) the increments in each of the compositions/axes to be enumerated}
}
\description{
Create a grid on a m-dim simplex, enumerating all points spaced by \code{step_size}
}
\details{
returns a matrix with \code{dim} columns and n rows that is increasingly large for increasing \code{dim}.
Below is a table of the row numbers for a \code{step_size} of 0.1 (default) and \code{dim} dimensions

\tabular{rrrr}{
  \code{dim} \tab \code{step_size} \tab n rows \tab time taken to run (sec) on i5-8400 \cr
  2 \tab 0.1 \tab 11 \tab 0.004 \cr
  3 \tab 0.1 \tab 66 \tab 0.039  \cr
  4 \tab 0.1 \tab 286 \tab 0.248  \cr
  5 \tab 0.1 \tab 1001 \tab 1.131  \cr
  6 \tab 0.1 \tab 3003 \tab 4.168  \cr
  7 \tab 0.1 \tab 8008 \tab 13.549  \cr
  8 \tab 0.1 \tab 19448 \tab  38.434 \cr
  9 \tab 0.1 \tab 43758 \tab 97.613
}
}
\examples{
mk_simplex_grid(2)
mk_simplex_grid(2, 0.05)
nrow(mk_simplex_grid(2))
system.time(mk_simplex_grid(2))

mk_simplex_grid(5, 0.5)
}
\author{
Ty Stanford <tystan@gmail.com>
}
