% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_three_comp.R
\name{plot_three_comp}
\alias{plot_three_comp}
\title{Plot 3-simplex data in a the triangle contrained 2-D plane (ternary diagram)}
\usage{
plot_three_comp(comp_data, col_var_nm = NULL, pt_alpha = 0.9)
}
\arguments{
\item{comp_data}{a column named  \code{data.frame} of \code{n} observations (rows). The first three columns are assumed the \code{3} compositional components (columns)}

\item{col_var_nm}{(default \code{NULL}) If not null and a scalar \code{character} name, a named column in \code{comp_data} corresponding to the points' values to be mapped a colour palette.}

\item{pt_alpha}{(default \code{0.9}) Transparency of the points plotted}
}
\description{
Plot 3-simplex data in a the triangle contrained 2-D plane (ternary diagram)
}
\details{
Plots a \code{ggplot} 2D scatterplot of the 3-simplex data (in a 2-D triangle).


Currently the optional \code{col} data are assumed \code{numeric} or 
\code{integer} values. Currently handling of factor and nominal 
variables is not implemented.
}
\examples{
(grid_3simplex <- (enumerate_simplex(p = 3, n = 10 - 3) + 1) / 10)
colnames(grid_3simplex) <- paste0("x", 1:ncol(grid_3simplex))
# simulate response variable based on 4-simplex ilrs
y <- simplex_to_ilr(grid_3simplex) \%*\% matrix(c(-1, 0.5), ncol = 1)
colnames(y) <- "outcome"
tern_dat <- as.data.frame(cbind(grid_3simplex, y))
plot_three_comp(tern_dat, col_var_nm = "outcome")
# No colour scale version:
# plot_three_comp(tern_dat)
}
\author{
Ty Stanford <tystan@gmail.com>
}
