% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_ilrs.R
\name{rotate_ilrs}
\alias{rotate_ilrs}
\title{Title}
\usage{
rotate_ilrs(ilr_dat, from_basis, to_basis)
}
\arguments{
\item{ilr_dat}{a \code{n} times \code{D-1} data.frame or matrix of ilrs using basis matrix \code{from_basis}}

\item{from_basis}{the \code{D} times \code{D-1} basis matrix used to generate \code{ilr_dat}}

\item{to_basis}{the \code{D} times \code{D-1} basis matrix wanting to transform \code{ilr_dat} to}
}
\value{
an equivalent ilr expression of \code{ilr_dat} using basis matrix \code{to_basis} (a \code{n} times \code{D-1} matrix of ilrs)
}
\description{
Title
}
\examples{
(comp_dat <- matrix(c(3, 3, 4, 3, 4, 3, 1, 2, 7, 1, 7, 2), byrow = TRUE, ncol = 3))
(b0 <- mk_vt(3, normalise = TRUE))
b1 <- b0
b1[, 1] <- -b1[, 1]
b1 <- -b1[3:1,]
b1
ilr_b0 <- log(comp_dat) \%*\% b0
ilr_b1 <- log(comp_dat) \%*\% b1
stopifnot(all(abs(ilr_b1 - rotate_ilrs(ilr_b0, b0, b1) < 1e-12)))
closure_rowwise(exp(rotate_ilrs(ilr_b0, b0, b1) \%*\% t(b1)), clo_val = 10)
}
