% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.R
\name{scale01}
\alias{scale01}
\title{Scale values in a vector to \code{[0, 1]} range}
\usage{
scale01(x, rm_na = TRUE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{rm_na}{(default \code{TRUE}) should \code{NA} values be removed (almost always a good idea)}
}
\description{
Scale values in a vector to \code{[0, 1]} range
}
\details{
Returns a vector of equal size of \code{x} with values scaled to the range [0,1]
based on the min and max values of \code{x}.
}
\examples{
set.seed(1234)
(x <- rnorm(10))
scale01(x)
plot(x, scale01(x))
}
\author{
Ty Stanford <tystan@gmail.com>
}
