% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_to_ilr.R
\name{simplex_to_ilr}
\alias{simplex_to_ilr}
\title{Convert \eqn{n \times D} simplex data to \eqn{n \times (D-1)} ilr-space data}
\usage{
simplex_to_ilr(simplex_dat)
}
\description{
Convert \eqn{n \times D} simplex data to \eqn{n \times (D-1)} ilr-space data
}
\examples{
grid_3simplex <- mk_simplex_grid(dim = 3, step_size = 0.1, rm_edges = TRUE) 
head(grid_3simplex)
# transform to 2-D ilr space using v = t(create_vt(3)) basis
grid_2ilr <- simplex_to_ilr(grid_3simplex) 
head(grid_2ilr)
plot(grid_2ilr[, 1], grid_2ilr[, 2], xlab = "ilr1", ylab = "ilr2")
}
