% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_comp_to_tetra.R
\name{trans_comp_to_tetra}
\alias{trans_comp_to_tetra}
\title{Transform 4-simplex data to a 3-D plotting coordinates}
\usage{
trans_comp_to_tetra(comp_dat, warn = FALSE)
}
\arguments{
\item{comp_dat}{a matrix of \code{n} observations (rows) and \code{4} compositional components (columns)}

\item{warn}{(default \code{FALSE}) The closure operation to 1 is applied to 
the observations in \code{comp_dat}, should a warning about pre-closure observations not being 1 be printed?}
}
\description{
Transform 4-simplex data to a 3-D plotting coordinates
}
\details{
Returns a \code{n x 4} \code{data.frame} of \code{(x, y, z)} 3-D plotting coordinates.
}
\examples{
(grid_4simplex <- mk_simplex_grid(4, 0.2, rm_edges = TRUE, nc = 1))
trans_comp_to_tetra(grid_4simplex)

}
\author{
Ty Stanford <tystan@gmail.com>
}
