% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clos_check.R
\name{clos_check}
\alias{clos_check}
\title{Check that rows sum up to the closure value}
\usage{
clos_check(x, clo_val = 1, warn = TRUE)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix} with of \code{n} observations (rows) and \code{D} compositional components (columns)}

\item{clo_val}{a positive closure value to check each row sums to this}

\item{warn}{(default \code{TRUE}) should a warning about pre-closure observations not being \code{clo_val} be printed?}
}
\description{
Check that rows sum up to the closure value
}
\examples{
(c_dat1 <- matrix(c(rep(0.25, 4), 1:4 / 10, 1:4), byrow = TRUE, ncol = 4))
clos_check(c_dat1)
(c_dat2 <- matrix(c(rep(0.25, 4), 1:4 / 10)     , byrow = TRUE, ncol = 4))
clos_check(c_dat2)
(c_dat3 <- matrix(c(rep(0.25, 4), 1:4 / 10, 1:4), byrow = TRUE, ncol = 4))
clos_check(c_dat3, clo_val = 10)
(c_dat4 <- 3 * matrix(c(rep(0.25, 4), 1:4 / 10), byrow = TRUE, ncol = 4))
clos_check(c_dat4, 3)


}
\author{
Ty Stanford <tystan@gmail.com>
}
