% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_to_power.R
\name{covar_to_power}
\alias{covar_to_power}
\title{Calculate the real valued power of the variance-covariance matrix}
\usage{
covar_to_power(covar_mat, pow = 1L, verbose = FALSE)
}
\arguments{
\item{covar_mat}{a positive-definite, square (n by n) matrix}

\item{pow}{the power to calculate}

\item{verbose}{provide working (FALSE, default)}
}
\value{
Used the eigen decomposition that

\code{covar_mat} \eqn{= U \Lambda U^T}{= U \Lambda U^T}

when 

\code{covar_mat} a positive-definite, square (n by n) matrix.

Therefore

\code{covar_mat ** pow} \eqn{= U \Lambda^{pow} U^T}{= U \Lambda^{pow} U^T}

Note 1: should also satisfy \code{(covar_mat ** pow) ** 1/pow = covar_mat}
Note 2: an implementation using  expm::power() or Matrix::chol() would 
probably be quicker/better
}
\description{
Calculate the real valued power of the variance-covariance matrix
}
\examples{
A <- diag(2)
covar_to_power(A, pow = 2, verbose = TRUE)
covar_to_power(A, pow = -1/2, verbose = TRUE)
covar_inverse_sqrt(A, verbose = TRUE)

B <- diag(c(1, 4))
covar_to_power(B, pow = 2, verbose = TRUE)
covar_to_power(B, pow = 1/2, verbose = TRUE)
covar_to_power(B, pow = -1/2, verbose = TRUE)
covar_inverse_sqrt(B, verbose = TRUE)

C <- matrix(c(4, -2, -2, 4), nrow = 2)
covar_to_power(C, pow = 2, verbose = TRUE)
covar_to_power(C, pow = 1/2, verbose = TRUE)
covar_to_power(C, pow = -1/2, verbose = TRUE)
covar_inverse_sqrt(C, verbose = TRUE)
covar_to_power(covar_to_power(C, pow = -1/2), pow = -2)
covar_to_power(covar_to_power(C, pow = 2), pow = 1/2)
covar_to_power(covar_to_power(C, pow = -1/3), pow = -3)
covar_to_power(covar_to_power(C, pow = +1/3), pow = +3)

}
