% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_simplex.R
\name{enumerate_simplex}
\alias{enumerate_simplex}
\title{Create a grid on a m-dim simplex}
\usage{
enumerate_simplex(p, n, verbose = TRUE)
}
\arguments{
\item{p}{the dimension of the simplex. This will be the number of columns in the returned data}

\item{n}{integer compositions/axes to be enumerated}
}
\description{
Create a grid on a m-dim simplex, enumerating all points spaced by \code{step_size}
}
\author{
Ty Stanford <tystan@gmail.com>
}
