% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_vt.R
\name{mk_vt}
\alias{mk_vt}
\title{Create the transpose of a V matrix (i.e., V^T)}
\usage{
mk_vt(n_comp, normalise = TRUE)
}
\arguments{
\item{n_comp}{Integer value \code{D} that is the number of compositional components (columns)}

\item{normalise}{do you want to normalise the columns of V^T (similarly, the rows of V) to unit vectors?}
}
\description{
Create the transpose of a V matrix (i.e., V^T)
}
\details{
Note that:

\code{ilr(x) = ln(x) \%*\% t(V)}

\code{-> x = clo(exp(ilr(x) \%*\% V))}

as

\code{t(V) \%*\% V =} \eqn{I_{D-1}}, and

\code{V \%*\% t(V) =} \eqn{I_{D} - \frac{1}{D} \times 1_{DxD}}
}
\examples{
mk_vt(4)
(VT <- mk_vt(3))
VT \%*\% t(VT) 
t(VT) \%*\% VT  
}
\author{
Ty Stanford <tystan@gmail.com>
}
