% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowProds.R
\name{rowProds}
\alias{rowProds}
\title{row _product_ version of rowSums()}
\usage{
rowProds(x)
}
\description{
row _product_ version of rowSums()
}
\examples{
rowProds(matrix(1:9, ncol = 3))
rowProds(matrix(1:9, ncol = 1))
rowProds(1:9)
}
